Fix auth page metadata export
